pyproject.toml
[tool.poetry]
name = "excel_agent_client"
version = "0.1.0"
description = ""
authors = ["Akimitsu Inoue <akimitsu.inoue@gmail.com>"]
readme = "README.md"
packages = [{include = "excel_agent_client"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
langchain-core = "^0.1.42"
langchain-openai = "^0.1.3"
langchain-community = "^0.0.32"
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
langchainhub = "^0.1.15"
nest-asyncio = "^1.6.0"
joblib = "^1.4.0"
tavily-python = "^0.3.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"
commitizen = "^3.6.0"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = false
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119
select = [
    "E",   # pycodestyle
]
ignore = ["D415"]

[tool.ruff.pydocstyle]
convention = "google"
